requirements.txt
python-dotenv
disnake
pymongo
sqlalchemy
importlib